Chapter9/email-filtering-agent/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1
python-multipart>=0.0.9
azure-storage-blob>=12.19.0
chromadb>=0.5.0
